requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.6.0

# Deep Learning
torch>=1.12.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Dashboard
streamlit>=1.20.0

# RL (Optional)
gymnasium>=0.28.0
stable-baselines3>=2.0.0

# Utilities
tqdm>=4.60.0
joblib>=1.1.0
pyyaml>=6.0
